Refactor: move the RewindStatus extension to its own file
